Use only secp256k1 operations in HD wallet derivation
Previously we BigUint for a single modulo addition operation
in get_child_key(). Replacing this will hopefully ensure that
we are closer to secure constant time operation.

That being said I'm pretty sure the Hmac + SHA2 libraries are not
constant time and already expose us to attacks in this area.