Remove failure
This patch removes the failure crate in exchange for many smaller
and more specific error types as appropriate in each module.

This might not have been the best use of time, failure may be
depricated but it's not exactly going to go away.

That being said this does remove a dependency and allow for more
detailed matching on specific errors in the future.