Fix bug: PrivateKey::from_secret() could panic
This bug was caused by PrivateKey::from_secret() relying on the
bigint library to produce fixed length 256 fit binary representations.

Obviously this isn't a constraint the library cares about and when it
finds a key with leading or trailing zeros they would not be printed
resulting in a panic on copy from slice.