Proto refactor, first pass
This is the first pass of an enormous refactor for deep_space to use
proto transactions.

The biggest decision here that needs the most justification is the one
to move our request library into deep_space itself. This was driven by
the need to import Prost for even basic message types. Once you do this
the entire Tonic rpc stack comes along as a transitive dependency.
Unless we go upstream and breakout the requests from the structs much
more carefully there's simply no advantage to having a small signing
library when it transitively pulls in the whole shebang anyways.

Other than that some API changes have been made to more explicitly
highlight if a node is syncing. The rest requires more testing.