Expand list of invalid log messages
This patch expands the list of invalid log messages for the raw_log
value in a tx response. Any of these key phrases will cause wait_for_tx
to halt immediately becuase it knows the tx has failed.

I'm toying with the idea of taking any non empty value of raw log as an
error, but I'm not totally sure that that's valid. While this method is
clunky it's garunteed not to fail a tx that should actually succeed.