Reduce code-duplication using send_tokens
Previously we re-implemented about 6 lines of message send logic
everywhere we sent a message. This was tenable when it was only a few
lines of boilerplate, but as we move onto doing more validation of
messages around fees and the response values the amount of boilerplate
is expanding.

For this reason we're moving every function to use send_tokens() which
will perform all real time message validation that is possible and we
can reduce duplication. We are also suggesting that all downstream
crates move to this same pattern to take advantage of the ever expanding
list of implemented validations.