Add high level send_message function
Previously Contact provided only send_transaction and send_tokens

These represent the very lowest level way to send a transaction and the
most abstracted way to send a transaction.

This commit adds send_message, an intermediate layer of abstraction that
handles many of the details around fees and transaction tracking without
requiring that a tx be fully formed and signed before it's passed in.

Also added are complete doc tests for these functions.