Recieve gzip on all grpc queries
This patch enables the recieve of gzip compressed data everywhere
that deep_space interacts with the Cosmos SDK grpc server. Ideally this
will result in lower bandwidth usage and more efficient sends of large
messages and large query responses.

Unfortunately the Cosmos SDK grpc server currently does not support gzip,
so what this patch currently does is mark to the server that we support
gzip and if in some future implementation gzip is supported compressed
responses may be sent.