Add gas awareness for get_fee_into()
This patch adds a get_params() function which can be used to query any
registered parameter. This is then used to query the maximum allowed gas
and bytes for a block, which can then be used in get_fee_info() to print
a helpful error when a transaction simulation shows that the transaction
simply can not be executed on a chain without a parameter change. By
adding an explicit error type here it's possible for consumer crates to
match on this error and take steps to reduce the gas use of a single
transaction, like unbundling messages.