Fix tx simulation in presence of a global fee
Althea L1 uses the x/evm module, which often requires the use of a
minimum global fee. Transaction simulation completely broke because of
this and led to even manually provided fee amounts to
contact.send_message() to fail in get_fee_info() and submit an empty
collection of fees. Adding an optional fee parameter to transaction
submission fixes this in the case where the caller provides a fee value.